replaced

+++ backward_analysis/src/main.rs
@@ -69,6 +69,7 @@ fn x_to_x_common<P>(
     factory: &PathFactory<P>,
     num_from: usize,
     num_processes: usize,
+    del_tmp_files: bool,
 ) -> anyhow::Result<()>
 where
     P: AsRef<Path>,
@@ -76,29 +77,37 @@ where
     let num_to = num_from + 1;
 
     // --- BackStep ---
+    println!("### PHASE: BACKSTEP ###");
     for num_doves in 2..=12 {
         std::fs::create_dir_all(dove_dir(factory.backstepped(num_to), num_doves))?;
     }
 
     for num_doves in 2..=12 {
+        println!("=== num_doves={num_doves} ===");
         let src_path = factory
             .num_dir(num_from)
             .join(format!("{num_doves:0>2}.tdl"));
         let dst_dir = factory.backstepped(num_to);
-        core_methods::backstep(src_path, dst_dir, num_doves, num_processes, 500_000_000)?;
+        core_methods::backstep(src_path, dst_dir, num_doves, num_processes, 300_000_000)?;
     }
 
     // --- Redistribute ---
+    println!("### PHASE: REDISTRIBUTE ###");
     for num_doves in 2..=12 {
+        println!("=== num_doves={num_doves} ===");
         let src_dir = dove_dir(factory.backstepped(num_to), num_doves);
         let dst_dir = dove_dir(factory.redistributed(num_to), num_doves);
         std::fs::create_dir_all(&dst_dir)?;
         core_methods::redistribute(src_dir, dst_dir, num_processes)?;
     }
-    std::fs::remove_dir_all(factory.backstepped(num_to))?;
+    if del_tmp_files {
+        std::fs::remove_dir_all(factory.backstepped(num_to))?;
+    }
 
-    // --- Trim ---
+    // --- Trim Simple ---
+    println!("### PHASE: TRIM SIMPLE ###");
     for num_doves in 2..=12 {
+        println!("=== num_doves={num_doves} ===");
         let src_dir = dove_dir(factory.redistributed(num_to), num_doves);
         let dst_dir = dove_dir(factory.trimmed_simply(num_to), num_doves);
         std::fs::create_dir_all(&dst_dir)?;
@@ -106,7 +115,9 @@ where
         core_methods::trim_simply(src_dir, dst_dir, win_paths, num_processes)?;
     }
 
-    std::fs::remove_dir_all(factory.redistributed(num_to))?;
+    if del_tmp_files {
+        std::fs::remove_dir_all(factory.redistributed(num_to))?;
+    }
     Ok(())
 }
 
@@ -114,6 +125,7 @@ fn win_to_lose<P>(
     factory: &PathFactory<P>,
     num_from: usize,
     num_processes: usize,
+    del_tmp_files: bool,
 ) -> anyhow::Result<()>
 where
     P: AsRef<Path>,
@@ -121,10 +133,12 @@ where
     let num_to = num_from + 1;
 
     // x_to_x_common
-    x_to_x_common(factory, num_from, num_processes)?;
+    x_to_x_common(factory, num_from, num_processes, del_tmp_files)?;
 
     // Trim Move
+    println!("### PHASE: TRIM MOVE ###");
     for num_doves in 2..=12 {
+        println!("=== num_doves={num_doves} ===");
         let src_dir = dove_dir(factory.trimmed_simply(num_to), num_doves);
         let dst_dir = dove_dir(factory.trimmed_move(num_to), num_doves);
         std::fs::create_dir_all(&dst_dir)?;
@@ -138,10 +152,14 @@ where
             num_processes,
         )?;
     }
-    std::fs::remove_dir_all(factory.trimmed_simply(num_to))?;
+    if del_tmp_files {
+        std::fs::remove_dir_all(factory.trimmed_simply(num_to))?;
+    }
 
     // Trim Put
+    println!("### PHASE: TRIM PUT ###");
     for num_doves in 2..=12 {
+        println!("=== num_doves={num_doves} ===");
         let src_dir = dove_dir(factory.trimmed_move(num_to), num_doves);
         let dst_dir = dove_dir(factory.trimmed_put(num_to), num_doves);
         std::fs::create_dir_all(&dst_dir)?;
@@ -168,10 +186,14 @@ where
             num_processes,
         )?;
     }
-    std::fs::remove_dir_all(factory.trimmed_move(num_to))?;
+    if del_tmp_files {
+        std::fs::remove_dir_all(factory.trimmed_move(num_to))?;
+    }
 
     // Trim Remove
+    println!("### PHASE: TRIM REMOVE ###");
     for num_doves in 2..=12 {
+        println!("=== num_doves={num_doves} ===");
         let src_dir = dove_dir(factory.trimmed_put(num_to), num_doves);
         let dst_dir = dove_dir(factory.trimmed_remove(num_to), num_doves);
         std::fs::create_dir_all(&dst_dir)?;
@@ -198,18 +220,23 @@ where
             num_processes,
         )?;
     }
-    std::fs::remove_dir_all(factory.trimmed_put(num_to))?;
+    if del_tmp_files {
+        std::fs::remove_dir_all(factory.trimmed_put(num_to))?;
+    }
 
     // Gather
+    println!("### PHASE: GATHER ###");
     std::fs::create_dir_all(factory.num_dir(num_to))?;
     for num_doves in 2..=12 {
+        println!("=== num_doves={num_doves} ===");
         core_methods::gather(
             dove_dir(factory.trimmed_remove(num_to), num_doves),
             factory.num_dir(num_to),
         )?;
     }
-    std::fs::remove_dir_all(factory.trimmed_remove(num_to))?;
-
+    if del_tmp_files {
+        std::fs::remove_dir_all(factory.num_tmp_dir(num_to))?;
+    }
     Ok(())
 }
 
@@ -217,6 +244,7 @@ fn lose_to_win<P>(
     factory: &PathFactory<P>,
     num_from: usize,
     num_processes: usize,
+    del_tmp_files: bool,
 ) -> anyhow::Result<()>
 where
     P: AsRef<Path>,
@@ -224,17 +252,22 @@ where
     let num_to = num_from + 1;
 
     // x_to_x_common
-    x_to_x_common(factory, num_from, num_processes)?;
+    x_to_x_common(factory, num_from, num_processes, del_tmp_files)?;
 
     // Gather
+    println!("### PHASE: GATHER ###");
     std::fs::create_dir_all(factory.num_dir(num_to))?;
     for num_doves in 2..=12 {
+        println!("=== num_doves={num_doves} ===");
         core_methods::gather(
             dove_dir(factory.trimmed_simply(num_to), num_doves),
-            factory.num_dir(num_to),
+            factory.num_dir(num_to).join(format!("{num_doves:0>2}.tdl")),
         )?;
     }
-    std::fs::remove_dir_all(factory.trimmed_simply(num_to))?;
+
+    if del_tmp_files {
+        std::fs::remove_dir_all(factory.num_tmp_dir(num_to))?;
+    }
     Ok(())
 }
 
@@ -242,13 +275,14 @@ fn advance_one_step(
     root: impl AsRef<Path>,
     num_from: usize,
     num_processes: usize,
+    del_tmp_files: bool,
 ) -> anyhow::Result<()> {
     let factory = PathFactory::new(root);
     match num_from {
         0 | 1 => return Err(anyhow::anyhow!("invalid num_from")),
         n => match n % 2 {
-            0 => lose_to_win(&factory, num_from, num_processes)?,
-            1 => win_to_lose(&factory, num_from, num_processes)?,
+            0 => lose_to_win(&factory, num_from, num_processes, del_tmp_files)?,
+            1 => win_to_lose(&factory, num_from, num_processes, del_tmp_files)?,
             _ => unreachable!(),
         },
     }
@@ -257,9 +291,9 @@ fn advance_one_step(
 }
 
 fn main() -> anyhow::Result<()> {
-    let root = PathBuf::from(r"...");
+    let root = PathBuf::from(r"..");
     let num_from = 2;
     let num_processes = 16;
-    advance_one_step(root, num_from, num_processes)?;
+    advance_one_step(root, num_from, num_processes, false)?;
     Ok(())
 }
